[refactor]: split logic into files
