+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/BooleanCalculator.spec.ts
@@ -1,4 +1,4 @@
-import { BooleanCalculator } from './';
+import { BooleanCalculator } from './BooleanCalculator';
 
 const booleanCalculator = new BooleanCalculator();
