+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/tokenHandlers/handleLogicToken.ts
@@ -1,0 +1,14 @@
+import { Logic, TokenType } from '../tokens';
+import { TokenHandler } from './tokenHandlers';
+
+export const handleLogicToken: TokenHandler = (context, action) => {
+  const newContext = { ...context };
+  const nextAllowedTokenTypes = [TokenType.Boolean, TokenType.Negation, TokenType.OpenGroup];
+
+  newContext.logic = action.token as Logic;
+
+  return {
+    context: newContext,
+    nextAllowedTokenTypes,
+  };
+};
