+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/tokenHandlers/handleNegationToken.ts
@@ -1,0 +1,14 @@
+import { TokenType } from '../tokens';
+import { TokenHandler } from './tokenHandlers';
+
+export const handleNegationToken: TokenHandler = (context, _action) => {
+  const nextAllowedTokenTypes = [TokenType.Boolean, TokenType.OpenGroup];
+  const newContext = { ...context };
+
+  newContext.shouldNegate = true;
+
+  return {
+    context: newContext,
+    nextAllowedTokenTypes,
+  };
+};
