+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/tokenHandlers/utils.ts
@@ -1,0 +1,6 @@
+export class LinkedListNode<T> {
+  next: LinkedListNode<T> | undefined;
+  previous: LinkedListNode<T> | undefined;
+
+  constructor(public value: T) {}
+}
