+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/tokenIterator.ts
@@ -1,0 +1,68 @@
+import { isGroupToken } from './utils';
+
+interface TokenIteratorValue {
+  token: string;
+  position: number;
+}
+
+export function getTokenIterator(booleanExpression: string): Iterator<TokenIteratorValue> {
+  if (!booleanExpression) {
+    throw new Error('Boolean expression must not be empty');
+  }
+
+  let position = 0;
+
+  const geSpaceCountStartingFrom = (startPosition: number) => {
+    let count = 0;
+
+    for (let l = booleanExpression.length; startPosition < l; startPosition++) {
+      const character = booleanExpression[startPosition];
+
+      if (/\s/.test(character)) {
+        count++;
+      } else {
+        break;
+      }
+    }
+
+    return count;
+  };
+
+  return {
+    next() {
+      let currentToken = '';
+
+      for (let l = booleanExpression.length; position < l; position++) {
+        const character = booleanExpression[position];
+        const isSpace = /\s/.test(character);
+        const isEnd = position === l - 1;
+        const isCurrentGroupToken = isGroupToken(character);
+        const isNextGroupToken = isGroupToken(booleanExpression[position + 1]);
+        const isTokenComplete = isSpace || isEnd || isCurrentGroupToken || isNextGroupToken;
+
+        if (!isSpace) {
+          currentToken += character;
+        }
+
+        if (isTokenComplete) {
+          if (isCurrentGroupToken || isNextGroupToken) {
+            position++;
+          }
+
+          position += geSpaceCountStartingFrom(position);
+          const newIsEnd = position === l - 1;
+
+          return {
+            value: {
+              token: currentToken,
+              position,
+            },
+            done: newIsEnd,
+          };
+        }
+      }
+
+      throw new Error('Iterator has already been finished');
+    },
+  };
+}
