+++ 1_The_Physical/Guess_Point_Level_1_Stateless/3_Manage_Complexity/3_2_Boolean_Calculator/src/tokens.ts
@@ -1,0 +1,20 @@
+export const booleanTokens = ['TRUE', 'FALSE'];
+export const logicTokens = ['AND', 'OR'];
+export const negationToken = 'NOT';
+export const openGroupToken = '(';
+export const closeGroupToken = ')';
+
+export enum Logic {
+  And = 'AND',
+  Or = 'OR',
+}
+
+export enum TokenType {
+  Boolean = 'Boolean',
+  // AND, OR
+  Logic = 'Logic',
+  Negation = 'Not',
+  OpenGroup = 'OpenGroup',
+  CloseGroup = 'CloseGroup',
+  Unknown = 'Unknown',
+}
